\name{dlmNu}
\alias{dlmNu}
\title{DLM Observation Disturbances}
\usage{
  dlmNu(y, m, mod)
}
\arguments{
  \item{y}{Data}

  \item{m}{Latent states}

  \item{mod}{An object of class \code{dlm}}
}
\value{
  \code{array}
}
\description{
  Returnes the estimated observation disturbances of a
  dynamic linear model.
}

